+++ src/Haskellorls/System/Posix/Files/ByteString.hs
@@ -1,13 +1,17 @@
 -- | "System.Posix.Files.ByteString" with extras.
 module Haskellorls.System.Posix.Files.ByteString
   ( destFileStatusRecursive,
+    major,
+    minor,
     module System.Posix.Files.ByteString,
   )
 where
 
 import Control.Monad.IO.Class
+import Data.Bits
 import System.FilePath.Posix.ByteString
 import System.Posix.Files.ByteString
+import System.Posix.Types (DeviceID)
 
 -- | A useful version of 'getFileStatus', to get a resolved destination file's
 -- status, but not a symbolic link self.
@@ -27,3 +31,11 @@ linkDestPath :: RawFilePath -> RawFilePath -> RawFilePath
 linkDestPath parPath linkPath
   | isAbsolute linkPath = linkPath
   | otherwise = takeDirectory parPath </> linkPath
+
+-- | A major number of a device.
+major :: DeviceID -> DeviceID
+major did = (did .&. 0x00000000000fff00) `shiftR` 8 .|. (did .&. 0xfffff00000000000) `shiftR` 32
+
+-- | A minor number of a device.
+minor :: DeviceID -> DeviceID
+minor did = (did .&. 0x00000000000000ff) `shiftR` 0 .|. (did .&. 0x00000ffffff00000) `shiftR` 12